requirements.txt
fastapi==0.116.1
groq==0.31.0
motor==3.7.1
openai==1.102.0
Pillow==11.3.0
pydantic==2.11.7
pydantic-settings==2.6.1
python-dotenv==1.1.1
Requests==2.32.5
torch==2.8.0
transformers==4.56.0
yagmail==0.15.293
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
passlib[bcrypt]==1.7.4
httpx==0.27.2
authlib==1.3.1
open-clip-torch
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
email-validator==2.3.0
